backend/requirements.txt
# Core dependencies
setuptools>=65.0.0
wheel>=0.38.0

# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0

# Vector database and embeddings
# Note: numpy will be installed as a dependency of other packages
# Using a version with pre-built wheels for Windows compatibility
chromadb>=0.4.18
sentence-transformers>=2.2.2

# LLM clients
groq>=0.4.1
google-generativeai>=0.3.1

# Document processing
pypdf>=3.17.4
python-docx>=1.1.0

# Utilities
tiktoken>=0.5.1

